feat: created reusable component to display added micronutrients to premix
