+++ src/app/pages/quickMaps/pages/costEffectiveness/interventionReview.module.ts
@@ -36,6 +36,7 @@ import { ReusableCostGraphComponent } from './interventionReview/utilities/reusa
 import { ReusableCostTableComponent } from './interventionReview/utilities/reusableCostTable/reusableCostTable.component';
 import { InterventionStepDetailsComponent } from './interventionReview/utilities/interventionStepDetails/interventionStepDetails.component';
 import { ComponentsModule } from 'src/app/components/components.module';
+import { MicroNutrientsInPremixTableComponent } from './interventionReview/utilities/microNutrientsInPremixTable/microNutrientsInPremixTable.component';
 
 @NgModule({
   declarations: [
@@ -59,6 +60,7 @@ import { ComponentsModule } from 'src/app/components/components.module';
     ReusableCostGraphComponent,
     ReusableCostTableComponent,
     InterventionStepDetailsComponent,
+    MicroNutrientsInPremixTableComponent,
   ],
   imports: [
     CommonModule,
