+++ src/app/pages/quickMaps/pages/costEffectiveness/interventionReview/pages/interventionBaseline/interventionBaseline.component.html
@@ -151,53 +151,8 @@
 
     </div>
 
+    <app-micro-nutrients-in-premix-table></app-micro-nutrients-in-premix-table>
 
-
-    <table mat-table [dataSource]="addedMnDatasource">
-
-      <ng-container matColumnDef="micronutrient">
-        <th class="row-title" mat-header-cell *matHeaderCellDef> Micronutrient </th>
-        <td class="column-title" mat-cell *matCellDef="let element">
-          <span>{{element.micronutrient}}</span>
-        </td>
-      </ng-container>
-      <ng-container matColumnDef="compound">
-        <th class="row-title" mat-header-cell *matHeaderCellDef> Compound </th>
-        <td class="column-title" mat-cell *matCellDef="let element">
-          <div class="double-cell">
-            <mat-form-field appearance="outline">
-              <mat-select #select matNativeControl>
-                <mat-option *ngFor="let item of element.compounds" [value]="item">
-                  {{item.compound}}
-                </mat-option>
-              </mat-select>
-            </mat-form-field>
-            <input type="number" matInput required [value]="select.value?.targetVal">
-            <!-- <span>{{select.value?.targetVal | json}}</span> -->
-          </div>
-        </td>
-      </ng-container>
-      <ng-container matColumnDef="targetVal">
-        <th class="row-title" mat-header-cell *matHeaderCellDef> Standard/target nutrient level at point of
-          fortification (mg/kg) </th>
-        <td class="column-title" mat-cell *matCellDef="let element">
-          <div class="double-cell">
-            <input type="number" matInput required [value]="selectedCompound?.targetVal">
-            <span>
-              <button mat-icon-button color="primary" (click)="inputType()">
-                {{buttonValue?"GFDx":"User Input"}}<mat-icon>arrow_right_alt</mat-icon>
-              </button>
-            </span>
-          </div>
-        </td>
-      </ng-container>
-      <tr mat-header-row *matHeaderRowDef="addedFVdisplayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: addedFVdisplayedColumns;"></tr>
-    </table>
-
-    <button class="item-reset" mat-flat-button (click)="addMN()">
-      <mat-icon class="red-icon">add</mat-icon> Add micronutrient
-    </button>
     <app-intervention-step-details [resetDefaultValues]="true" [showFocusMn]="true" [showUserValues]="true"
       [showMnUnits]="true"></app-intervention-step-details>
